msm: mdss: allow unused SMP for targets with high number of SMPs
Currently MDP driver rejects reusing of a pipe if this reuse results
in wastage of SMPs. i.e. VIG0 pipe reused from YUV layer to RGB layer.
In this case, since RGB layer needs only plane0, plane1 SMPs are unused.
This kind of wasting leads to un-necessary GPU fall-backs for composition
when number of SMPs and number of pipes are less. For this reason driver
used to prohibit such configuration. But in high tier chip-sets, there
are enough number of SMPs and pipes are available such that these
transient wastage of SMPs can be absorbed without much impact.

Change-Id: Ia7baec801add81e5c97bded920ae5bdca5083e0b
Signed-off-by: Ujwal Patel <ujwalp@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>
Signed-off-by: engstk <eng.stk@sapo.pt>